shared/vpn-plugin-utils: load the editor from the same place as plugin
If passed a relative path, load the editor .so from the same directory
as the plugin .so. This is useful for development, as it allows running
the editor plugin from the build tree conveniently.

NetworkManager/NetworkManager#242

(imported from NetworkManager commit 25f625b4fc58f56b7ab84dbcfc155bc210f3ee66)