Improving handling of snoop of MPPE keys
Also adding extra code for handling of EAP protocol. Will steal the MPPE keys
if EAP-MSCHAPv2 or EAP-TLS is in use. However, it looks like WIN2K16 require
the HLAK to be generated and sent during the auth-stage, and if delayed to
ip-up stage then the MPPE keys are expected to be all zero'd out.