Restore the subject name from VPN configuration
Without it, the second time you open and save the VPN configuration causes
the nm-sstp-service to omit the "name <identity>" and thus the EAP-TLS
negotiation fails.

This value is automatically obtained by the subject name of the certificate
and can be overriden in the Advanced settings dialog by setting a value for
Identiy (see TLS page).

Like for any instance of EAP, the Identity is used on the Authentication
server to propose the correct EAP authentication. It is also used to establish
the correct authentication credentials in case of EAP-TLS, i.e. the identity
as presented to the server during EAP Identity should match that of the
certificate's subject name.

Signed-off-by: Eivind Næss <eivnaes@yahoo.com>